Set SSG API call
